+++ packages/components/atoms/f-form-field/src/components/FormField.vue
@@ -11,49 +11,28 @@
         <div
             :class="$style['c-formField-fieldWrapper']">
             <form-label
-                v-if="!isInline"
-                :label-style="normalisedLabelStyle"
+                v-if="shouldShowLabelText"
                 :label-for="uniqueId"
-                :is-inline="isInline"
                 :is-disabled="isDisabled"
-                :label-details="labelDetails"
+                v-bind="$props"
                 :data-test-id="testId.label">
                 {{ labelText }}
-                <template #description>
-                    <span
-                        v-if="hasInputDescription"
-                        :class="[
-                            'u-spacingTop',
-                            'u-spacingBottom--large',
-                            $style['c-formField-label-description']
-                        ]">
-                        <slot />
-                    </span>
-                </template>
             </form-label>
 
             <form-field-affixed
                 v-if="isAffixedField"
                 :id="uniqueId"
                 :attributes="$attrs"
+                v-bind="$props"
                 :type="normalisedInputType"
-                :value="value"
-                :field-size="fieldSize"
-                :prefix="prefix"
-                :suffix="suffix"
-                :has-error="hasError"
                 v-on="listeners" />
 
             <form-dropdown
                 v-else-if="isDropdown"
                 :id="uniqueId"
                 :attributes="$attrs"
-                :type="normalisedInputType"
-                :value="value"
-                :field-size="fieldSize"
-                :has-error="hasError"
+                v-bind="$props"
                 :has-icon="hasLeadingIcon"
-                :dropdown-options="dropdownOptions"
                 v-on="listeners" />
 
             <textarea
@@ -90,16 +69,6 @@
                     }]"
                 v-on="listeners"
             >
-
-            <form-label
-                v-if="isInline"
-                :id="`label-${uniqueId}`"
-                :label-for="uniqueId"
-                :label-style="normalisedLabelStyle"
-                :is-inline="isInline"
-                :data-test-id="`${testId.label}--inline`">
-                {{ labelText }}
-            </form-label>
 
             <span
                 v-if="hasLeadingIcon"
@@ -145,6 +114,7 @@ import FormDropdown from './FormDropdown.vue';
 import FormLabel from './FormLabel.vue';
 import debounce from '../services/debounce';
 import tenantConfigs from '../tenants';
+
 import {
     CUSTOM_INPUT_TYPES,
     DEFAULT_INPUT_TYPE,
@@ -153,9 +123,7 @@ import {
     VALID_AFFIXED_INPUT_TYPES,
     VALID_TRAILING_ICON_INPUT_TYPES,
     DEFAULT_FIELD_SIZE,
-    VALID_FIELD_SIZES,
-    VALID_LABEL_STYLES,
-    MOBILE_WIDTH
+    VALID_FIELD_SIZES
 } from '../constants';
 
 export default {
@@ -186,10 +154,9 @@ export default {
             validator: value => ((VALID_INPUT_TYPES.indexOf(value) !== -1) || (CUSTOM_INPUT_TYPES.indexOf(value) !== -1))// The prop value must match one of the valid input types
         },
 
-        labelStyle: {
-            type: String,
-            default: 'default',
-            validator: value => (VALID_LABEL_STYLES.indexOf(value) !== -1) // The prop value must match one of the valid label types
+        shouldShowLabelText: {
+            type: Boolean,
+            default: true
         },
 
         fieldSize: {
@@ -228,9 +195,9 @@ export default {
             default: undefined
         },
 
-        hasInputDescription: {
-            type: Boolean,
-            default: false
+        labelDescription: {
+            type: String,
+            default: ''
         },
 
         labelDetails: {
@@ -270,13 +237,6 @@ export default {
             return DEFAULT_INPUT_TYPE;
         },
 
-        normalisedLabelStyle () {
-            if (VALID_LABEL_STYLES.includes(this.labelStyle)) {
-                return this.labelStyle;
-            }
-            return '';
-        },
-
         formFieldLocale () {
             return globalisationServices.getLocale(tenantConfigs, this.locale, this.$i18n);
         },
@@ -310,11 +270,6 @@ export default {
                 label: formFieldName ? `formfield-${formFieldName}-label` : 'formfield-label',
                 textarea: formFieldName ? `formfield-${formFieldName}-textarea` : 'formfield-textarea'
             };
-        },
-
-        isInline () {
-            return (this.windowWidth < MOBILE_WIDTH && this.labelStyle === 'inlineNarrow') ||
-                this.labelStyle === 'inline';
         },
 
         isDropdown () {
@@ -473,11 +428,6 @@ $form-input-iconSize                           : 18px;
         .c-formField-field {
             border-radius: $form-input-borderRadius $form-input-borderRadius 0 0;
         }
-    }
-
-    .c-formField-label-description {
-        display: block;
-        font-weight: normal;
     }
 
     .c-formField-icon {